+++ include/eosio.wps/eosio.wps.hpp
@@ -407,14 +407,14 @@ public:
      *
      * - `{name} proposer` - proposer
      * - `{name} proposal_name` - proposal name
-     * - `{bool} [next=false]` - (true/false) activate proposal at the next voting period
+     * - `{bool} voting_period` - activate proposal at the specified voting period (must be current or next)
      *
      * ```bash
-     * cleos push action eosio.wps activate '["myaccount", "mywps", false]' -p myaccount
+     * cleos push action eosio.wps activate '["myaccount", "mywps", "2019-11-25T00:00:00"]' -p myaccount
      * ```
      */
     [[eosio::action]]
-    void activate( const eosio::name proposer, const eosio::name proposal_name, const bool next );
+    void activate( const eosio::name proposer, const eosio::name proposal_name, const eosio::time_point_sec voting_period );
 
     /**
      * ## ACTION `refund`
@@ -560,6 +560,24 @@ public:
     void setparams( const wps_parameters params );
 
     /**
+     * ## ACTION `complete`
+     *
+     * Complete WPS voting period
+     *
+     * - authority: `any`
+     *
+     * ### params
+     *
+     * - `{time_point_sec} voting_period` - voting period to complete
+     *
+     * ```bash
+     * cleos push action eosio.wps complete '["2019-11-25T00:00:00"]' -p eosio.wps
+     * ```
+     */
+    [[eosio::action]]
+    void complete( const eosio::time_point_sec voting_period );
+
+    /**
      * TESTING ONLY
      *
      * Should be removed in production
@@ -567,6 +585,11 @@ public:
     [[eosio::action]]
     void clean( const eosio::name table, const std::optional<eosio::name> scope );
 
+    /**
+     * TESTING ONLY
+     *
+     * Should be removed in production
+     */
     [[eosio::action]]
     void setstate( const state_row params );
 
@@ -605,6 +628,7 @@ private:
     void proposal_to_periods( const eosio::name proposal_name, const eosio::time_point_sec start_voting_period, const uint8_t duration, const eosio::name ram_payer );
     eosio::checksum256 get_tx_id();
     void activate_proposal( const eosio::name proposer, const eosio::name proposal_name, const eosio::time_point_sec start_voting_period );
+    void send_deferred( const eosio::action action, const uint64_t key, const uint64_t interval );
 
     // transfers
     void add_transfer( const eosio::name type, const eosio::name from, const eosio::name to, const eosio::asset quantity, const eosio::string memo );