+++ src/eosio.wps.cpp
@@ -1,13 +1,21 @@
 #include <eosio.wps/eosio.wps.hpp>
 
-#include "clean.cpp"
+// main
 #include "drafts.cpp"
 #include "on_notify.cpp"
-#include "proposals.cpp"
+#include "activate.cpp"
 #include "deposits.cpp"
 #include "periods.cpp"
 #include "settings.cpp"
 #include "vote.cpp"
 #include "proposers.cpp"
-#include "get_tx_id.cpp"
-#include "transfers.cpp"
+#include "complete.cpp"
+
+// utils
+#include "utils/get_tx_id.cpp"
+#include "utils/deferred.cpp"
+
+// testing
+#include "testing/transfers.cpp"
+#include "testing/setstate.cpp"
+#include "testing/clean.cpp"