+++ src/settings.cpp
@@ -28,15 +28,20 @@ void wps::setparams( const wps_parameters params )
     _settings.set( params, ram_payer );
 }
 
-/**
- * TESTING ONLY
- *
- * Should be removed in production
- */
-void wps::setstate( const state_row params )
+void wps::add_funding( const eosio::asset quantity )
 {
-    require_auth( get_self() );
     const eosio::name ram_payer = get_self();
 
-    _state.set( params, ram_payer );
+    auto state = _state.get_or_default();
+    state.available_funding += quantity;
+    _state.set( state, ram_payer );
+}
+
+void wps::sub_funding( const eosio::asset quantity )
+{
+    const eosio::name ram_payer = get_self();
+
+    auto state = _state.get_or_default();
+    state.available_funding -= quantity;
+    _state.set( state, ram_payer );
 }